+++ src/todoist/request.rs
@@ -113,11 +113,11 @@ fn maybe_start_spinner(config: &Config) -> Option<Spinner> {
     match env::var("DISABLE_SPINNER") {
         Ok(_) => None,
         _ => {
-            if let Some(true) = config.spinners {
+            if let Some(false) = config.spinners {
+                None
+            } else {
                 let sp = Spinner::new(SPINNER, MESSAGE.into());
                 Some(sp)
-            } else {
-                None
             }
         }
     }
@@ -135,5 +135,35 @@ pub fn new_uuid() -> String {
         String::from(FAKE_UUID)
     } else {
         Uuid::new_v4().to_string()
+    }
+}
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::test;
+
+    #[test]
+    fn test_maybe_start_spinner() {
+        let config = test::fixtures::config();
+
+        // true spinner
+        let response = maybe_start_spinner(&config);
+        assert!(matches!(response, Some(Spinner { .. })));
+
+        // false spinner
+        let config = Config {
+            spinners: Some(false),
+            ..config
+        };
+        let response = maybe_start_spinner(&config);
+        assert!(matches!(response, None));
+
+        // null spinner
+        let config = Config {
+            spinners: None,
+            ..config
+        };
+        let response = maybe_start_spinner(&config);
+        assert!(matches!(response, Some(Spinner { .. })));
     }
 }